Add ‘assessment’ controller logic
