+++ linnote/templates/assessments.html
@@ -12,7 +12,7 @@
 <ul>
     {% for assessment in assessments %}
     <li>
-        <a href="{{ url_for('assessment', name=assessment.stem) }}">
+        <a href="#">
             {{ assessment.stem }}
         </a>
     </li>
