+++ requirements.txt
@@ -3,4 +3,3 @@ matplotlib==2.1.0
 pandas==0.21.0
 xlrd==1.1.0
 flask==0.12.2
-wtforms==2.1
